Offspec and balanced roles feature
Added the idea of "offspec" roles, and some logic to fill "main spec" roles with those offspec people if we don't have enough main spec tanks or healers to fill the groups.

Also added explicit commands for testing instead of using a global debug variable.

Added the idea of a "greedy" wheel algorithm where it always tries to fill tanks and healers in all groups even if they aren't full of other guildmates. This can cause people to have to play their offspec in a pug group.