Add offspec pulling prioritization
Refactored functions to make this not just one giant block of code.

Added more debugging local print statements.

Improved the logic for moving people around to fill groups. It not follows a logical path as if we were manually making groups inside of WoW. It's not algorithmically optimal, but it's much easier to follow and will be easier to modify in the future.